fix overflowing buffer made worse by 0d26e6f or possibly created by that commit. fixes #30 better.
the debugger was replacing comments and names inside a buffer returned from the old portable disassembly function, which was returning a privately retained 64 byte buffer. previously names were limited to 30 characters, so it's possible expanding names in disassembly never blew that buffer.

testing of the 0d26e6f commit either stressed it harder than ever before, or hard enough to break it for the first time

solved by copying the disassembly buffer into a new buffer for expanding address labels to names, etc