split out some of the "gameful" callbacks into dedicated functions
My personal favorite is the WM_MOUSEWHEEL one. The lParam contains cursor X and
Y coords, but someone had misinterpreted it as a handle to the scrollbar and
used it to update the scroll info. So what happened? Garbage pointer
dereference leading to nonsense scrolling and crashes? Nope! The ScrollInfo
calls silently did nothing and everything worked as intended.