revive that "dead" EN_CHANGE code in a new callback
I have absolutely no idea how the old version ever worked!
Seriously, it makes no sense. It checked if the HIWORD of wParam equaled
EN_CHANGE even though it was in a BN_CLICKED case...but it still got the events.
This implies that BN_CLICKED == EN_CHANGE, but 0 very much does not equal 0x300.