down with hotkeys, long live accelerators!
Hotkeys are SUPER jank. There should be some kind of warning in the docs or
something. They are an operating system-level message, and if consumed, NO OTHER
APPLICATION GETS TO HAVE IT. So this Ctrl+A hotkey meant the hex editor couldn't
use it anymore. Neither could Visual Studio. Or Notepad++. Or Chrome.
That's BAD.